MOD issue

+++ template.cpp
@@ -67,7 +67,7 @@
 #define fin freopen("in.txt", "r", stdin)
 #define fout freopen("out.txt", "w", stdout)
 #define fast ios_base::sync_with_stdio(false); cin.tie(NULL)
-#define MOD 1e9+7
+#define MOD (ll)1e9+7
 #define gamma 0.57721566490153286060651209008240243
 #define PI acos(-1)
 #define pi 2*acos(0)
